Added error handling, page reload after upload media, client side image resizing

+++ server/users/media/upload/upload.js
@@ -8,6 +8,8 @@ var formidable = require('formidable');
 var tmpDir = require('os').tmpdir();
 var fs = require('fs');
 var async = require('async');
+var numCPUs = require('os').cpus().length;
+
 
 module.exports = function (N, apiPath) {
   N.validate(apiPath, {
@@ -20,6 +22,31 @@ module.exports = function (N, apiPath) {
       format: 'mongo',
       required: true
     }
+  });
+
+
+  // Delete files in array
+  //
+  var unlinkFiles = function (files, callback) {
+    async.each(files, function (file, next) {
+      fs.unlink(file.path, next);
+    }, callback);
+  };
+
+
+  // Delete medias from DB
+  //
+  var removeMedias = function (medias, callback) {
+    async.each(medias, function (media, next) {
+      media.remove(next);
+    }, callback);
+  };
+
+
+  // Fetch owner by hid
+  //
+  N.wire.before(apiPath, function fetch_user_by_hid(env, callback) {
+    N.wire.emit('internal:users.fetch_user_by_hid', env, callback);
   });
 
 
@@ -49,13 +76,14 @@ module.exports = function (N, apiPath) {
     // TODO: check quota
     // TODO: check permissions
 
-    if (env.user_info.is_guest) {
-      callback(N.io.NOT_AUTHORIZED);
+    // Wrong member_hid parameter
+    if (env.data.user._id.toString() !== env.data.album.user_id.toString()) {
+      callback(N.io.NOT_FOUND);
       return;
     }
 
     // Check is current user owner of album
-    if (env.session.user_id.toString() !== env.data.album.user_id.toString()) {
+    if (env.user_info.is_guest || env.session.user_id.toString() !== env.data.album.user_id.toString()) {
       callback(N.io.NOT_AUTHORIZED);
       return;
     }
@@ -72,28 +100,38 @@ module.exports = function (N, apiPath) {
     var form = new formidable.IncomingForm();
 
     form.uploadDir = tmpDir;
-    form.on('fileBegin', function (name, file) {
+    form.on('file', function (name, file) {
+      // Check file type
+      if (N.config.options.users.media_uploads.allowed_types.indexOf(file.type) < 0) {
+        // Type incorrect - skip file
+        fs.unlink(file.path);
+        return;
+      }
 
-      // TODO: check file.type and unlink wrong
+      // Check file size
+      if (N.config.options.users.media_uploads.max_size_kb < (file.size / 1024)) { // file.size in bytes
+        // Size incorrect - skip file
+        fs.unlink(file.path);
+        return;
+      }
+
       files.push(file);
 
     }).on('aborted', function () {
 
-      env.data.files = files;
-      env.data.success = false;
+      env.data.upload_files = files;
+      env.data.upload_success = false;
       callback();
 
     }).on('error', function (e) {
 
-      files.forEach(function (file) {
-        fs.unlink(file.path);
+      unlinkFiles(files, function () {
+        callback(e);
       });
-      callback(e);
-
     }).on('end', function () {
 
-      env.data.files = files;
-      env.data.success = true;
+      env.data.upload_files = files;
+      env.data.upload_success = true;
       callback();
 
     }).parse(env.origin.req);
@@ -106,20 +144,16 @@ module.exports = function (N, apiPath) {
   N.wire.after(apiPath, function prepare_media(env, callback) {
     var Media = N.models.users.Media;
     var Album = N.models.users.Album;
-    var files = env.data.files;
-    console.log(files);
+    var files = env.data.upload_files;
+    var medias = [];
 
     // User abort
-    if (!env.data.success) {
-      files.forEach(function (file) {
-        fs.unlink(file.path);
-      });
-
-      callback();
+    if (!env.data.upload_success) {
+      unlinkFiles(files, callback);
       return;
     }
 
-    async.eachSeries(files, function (file, next) {
+    async.eachLimit(files, numCPUs, function (file, next) {
       Media.createImage(file.path, function (err, fileId) {
         if (err) { return next(err); }
 
@@ -128,25 +162,34 @@ module.exports = function (N, apiPath) {
         media.album_id = env.data.album._id;
         media.created_at = new Date();
         media.file_id = fileId;
-        media.save(next);
+        media.save(function (err) {
+          if (err) { return next(err); }
+
+          medias.push(media);
+          next();
+        });
       });
     }, function (err) {
-      if (err) {
-        files.forEach(function (file) {
-          fs.unlink(file.path);
+      unlinkFiles(files, function () {
+        if (err) {
+          // Try to clean up dirty data on error
+          removeMedias(medias, function () {
+            callback(err);
+          });
+          return;
+        }
+
+        Album.updateInfo(env.data.album._id, function (err) {
+          if (err) {
+            // Try to clean up dirty data on error
+            removeMedias(medias, function () {
+              callback(err);
+            });
+            return;
+          }
+
+          callback();
         });
-        callback(err);
-        return;
-      }
-
-      Album.updateInfo(env.data.album._id, function () {
-        if (err) { return callback(err); }
-
-        files.forEach(function (file) {
-          fs.unlink(file.path);
-        });
-
-        callback();
       });
     });
   });
